move utils to lib dir
