--- tact_project/mapper/mapper_utils.py
@@ -1,8 +1,0 @@
-from decimal import Decimal
-
-def str_to_decimal(a_str):
-    try:
-        return Decimal(a_str.replace(',','').replace('$', ''))
-    except:
-        return 0
-
